worker/requirements.txt
fastapi
uvicorn
python-multipart
requests
pymongo
sentence-transformers
faiss-cpu
pdfplumber
pytesseract
Pillow
python-dotenv
